Relax reschedule state (#28)
* Relax state reschedule restriction

This changes reschedule to allow setting/unsetting of state which was
previously not supported.

This change does NOT change state being ignored by enqueue. This is a
concious decision to avoid tight-coupling and abstraction leakage. The
Worker/Consumer of Jobs sets and owns the state which the Enqueuers need
know nothing about.